+++ R/TreatmentPatterns-package.R
@@ -15,9 +15,6 @@
 #' @importFrom htmlwidgets JS
 #' @importFrom stats sd median
 #' @importFrom tidyr pivot_wider
-#' @importFrom googleVis gvisSankey
-#' @importFrom rjson toJSON
-#' @importFrom stringi stri_replace_all_fixed
 ## usethis namespace: end
 NULL